Move De-Identifier Functions into own Module
